Update conda env creation to use mache
To do this, I am now creating a temporary conda environment,
`temp_conda_install`, for bootstrapping.  This way, we don't need
to install `mache`, `progressbar2` and `jinja2` into the base
environment.  The temporary environment gets deleted after the
compass environment has been created.
